Move setMapTriggerEpic to epics/mapInfo
