+++ web/client/plugins/Map.jsx
@@ -426,5 +426,8 @@ module.exports = {
         maptype: require('../reducers/maptype'),
         additionallayers: require('../reducers/additionallayers')
     },
-    epics: require('../epics/map')
+    epics: {
+        ...require('../epics/map'),
+        ...require('../epics/mapInfo')
+    }
 };
